+++ labbook_BE/default_settings.py
@@ -1,6 +1,6 @@
 # -*- coding:utf-8 -*-
 
-APP_VERSION = '2.9.0-beta.6'
+APP_VERSION = '2.9.0-rc.1'
 
 # ######################################
 #  Database parameters