+++ labbook_FE/default_settings.py
@@ -1,3 +1,3 @@
 # -*- coding:utf-8 -*-
 
-APP_VERSION = '2.9.0-beta.6'
+APP_VERSION = '2.9.0-beta.7'